feat: 콘솔앱 model 파트 추가 및 폴더 정리
