--- BankManagerConsoleApp/BankManagerConsoleApp/main.swift
@@ -1,9 +1,0 @@
-//
-//  BankManagerConsoleApp - main.swift
-//  Created by yagom. 
-//  Copyright © yagom academy. All rights reserved.
-// 
-
-var bank: Bank = Bank()
-
-bank.openSystem()
