+++ BankManagerConsoleApp/Model/BankWork.swift
