+++ BankManagerConsoleApp/Model/BankWorker.swift
