+++ BankManagerConsoleApp/Model/DataStruct/LinkedList.swift
