+++ BankManagerConsoleApp/Model/DataStruct/Node.swift
