+++ BankManagerUIApp/BankManagerUIApp/Resources/AppDelegate.swift
