+++ BankManagerUIApp/BankManagerUIApp/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
