+++ BankManagerUIApp/BankManagerUIApp/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
