+++ BankManagerUIApp/BankManagerUIApp/Resources/Assets.xcassets/Contents.json
