+++ BankManagerUIApp/BankManagerUIApp/Resources/SceneDelegate.swift
